app/requirements.txt
Django
psycopg2-binary
python-docx